refactor: 예외 객체 패키지 분리 및 이동
