+++ src/main/java/roomescape/admin/application/exception/NotAdminException.java
@@ -1,4 +1,4 @@
-package roomescape.global.exception;
+package roomescape.admin.application.exception;
 
 public class NotAdminException extends RuntimeException {
     public NotAdminException(String message) {
