+++ src/main/java/roomescape/auth/application/exception/InvalidMemberException.java
@@ -1,11 +1,12 @@
-package roomescape.global.exception;
+package roomescape.auth.application.exception;
 
 import org.springframework.http.HttpStatus;
 
-public class InvalidTokenException extends RuntimeException {
+public class InvalidMemberException extends RuntimeException {
+
     private final HttpStatus statusCode;
 
-    public InvalidTokenException(HttpStatus statusCode, String message) {
+    public InvalidMemberException(String message, HttpStatus statusCode) {
         super(message);
         this.statusCode = statusCode;
     }
