--- src/main/java/roomescape/global/exception/DeleteReservationException.java
@@ -1,7 +1,0 @@
-package roomescape.global.exception;
-
-public class DeleteReservationException extends RuntimeException{
-    public DeleteReservationException(String message) {
-        super(message);
-    }
-}
