--- src/main/java/roomescape/global/exception/DeleteThemeException.java
@@ -1,7 +1,0 @@
-package roomescape.global.exception;
-
-public class DeleteThemeException extends RuntimeException{
-    public DeleteThemeException(String message) {
-        super(message);
-    }
-}
