--- src/main/java/roomescape/global/exception/DeleteTimeException.java
@@ -1,7 +1,0 @@
-package roomescape.global.exception;
-
-public class DeleteTimeException extends RuntimeException{
-    public DeleteTimeException(String message) {
-        super(message);
-    }
-}
