--- src/main/java/roomescape/global/exception/DuplicateReservationException.java
@@ -1,7 +1,0 @@
-package roomescape.global.exception;
-
-public class DuplicateReservationException extends RuntimeException{
-    public DuplicateReservationException(String message) {
-        super(message);
-    }
-}
