--- src/main/java/roomescape/global/exception/DuplicateTimeException.java
@@ -1,7 +1,0 @@
-package roomescape.global.exception;
-
-public class DuplicateTimeException extends RuntimeException{
-    public DuplicateTimeException(String message) {
-        super(message);
-    }
-}
