--- src/main/java/roomescape/global/exception/GetThemeException.java
@@ -1,7 +1,0 @@
-package roomescape.global.exception;
-
-public class GetThemeException extends RuntimeException{
-    public GetThemeException(String message) {
-        super(message);
-    }
-}
