--- src/main/java/roomescape/global/exception/GetTimeException.java
@@ -1,7 +1,0 @@
-package roomescape.global.exception;
-
-public class GetTimeException extends RuntimeException{
-    public GetTimeException(String message) {
-        super(message);
-    }
-}
