--- src/main/java/roomescape/global/exception/PastTimeException.java
@@ -1,7 +1,0 @@
-package roomescape.global.exception;
-
-public class PastTimeException extends RuntimeException{
-    public PastTimeException(String message) {
-        super(message);
-    }
-}
