+++ src/main/java/roomescape/reservation/application/exception/DeleteReservationException.java
@@ -1,0 +1,7 @@
+package roomescape.reservation.application.exception;
+
+public class DeleteReservationException extends RuntimeException {
+    public DeleteReservationException(String message) {
+        super(message);
+    }
+}
