+++ src/main/java/roomescape/reservation/application/exception/DeleteThemeException.java
@@ -1,0 +1,7 @@
+package roomescape.reservation.application.exception;
+
+public class DeleteThemeException extends RuntimeException {
+    public DeleteThemeException(String message) {
+        super(message);
+    }
+}
