+++ src/main/java/roomescape/reservation/application/exception/DeleteTimeException.java
@@ -1,0 +1,7 @@
+package roomescape.reservation.application.exception;
+
+public class DeleteTimeException extends RuntimeException {
+    public DeleteTimeException(String message) {
+        super(message);
+    }
+}
