+++ src/main/java/roomescape/reservation/application/exception/DuplicateReservationException.java
@@ -1,0 +1,7 @@
+package roomescape.reservation.application.exception;
+
+public class DuplicateReservationException extends RuntimeException {
+    public DuplicateReservationException(String message) {
+        super(message);
+    }
+}
