+++ src/main/java/roomescape/reservation/application/exception/DuplicateTimeException.java
@@ -1,0 +1,7 @@
+package roomescape.reservation.application.exception;
+
+public class DuplicateTimeException extends RuntimeException {
+    public DuplicateTimeException(String message) {
+        super(message);
+    }
+}
