+++ src/main/java/roomescape/reservation/application/exception/GetThemeException.java
@@ -1,0 +1,7 @@
+package roomescape.reservation.application.exception;
+
+public class GetThemeException extends RuntimeException {
+    public GetThemeException(String message) {
+        super(message);
+    }
+}
