+++ src/main/java/roomescape/reservation/application/exception/GetTimeException.java
@@ -1,0 +1,7 @@
+package roomescape.reservation.application.exception;
+
+public class GetTimeException extends RuntimeException {
+    public GetTimeException(String message) {
+        super(message);
+    }
+}
