+++ src/main/java/roomescape/reservation/application/exception/PastTimeException.java
@@ -1,0 +1,7 @@
+package roomescape.reservation.application.exception;
+
+public class PastTimeException extends RuntimeException {
+    public PastTimeException(String message) {
+        super(message);
+    }
+}
