+++ src/main/java/roomescape/reservation/application/service/ReservationService.java
@@ -6,14 +6,14 @@ import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import roomescape.admin.presentation.dto.AdminReservationRequest;
-import roomescape.global.exception.DuplicateReservationException;
-import roomescape.global.exception.GetThemeException;
-import roomescape.global.exception.GetTimeException;
-import roomescape.global.exception.InvalidMemberException;
-import roomescape.global.exception.PastTimeException;
+import roomescape.auth.application.exception.InvalidMemberException;
 import roomescape.member.application.repository.MemberRepository;
 import roomescape.member.domain.Member;
 import roomescape.reservation.application.dto.CreateReservationRequest;
+import roomescape.reservation.application.exception.DuplicateReservationException;
+import roomescape.reservation.application.exception.GetThemeException;
+import roomescape.reservation.application.exception.GetTimeException;
+import roomescape.reservation.application.exception.PastTimeException;
 import roomescape.reservation.application.repository.ReservationRepository;
 import roomescape.reservation.application.repository.ReservationTimeRepository;
 import roomescape.reservation.application.repository.ThemeRepository;
