+++ src/main/java/roomescape/reservation/application/service/ReservationTimeService.java
@@ -3,12 +3,11 @@ package roomescape.reservation.application.service;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.global.exception.DeleteTimeException;
-import roomescape.global.exception.DuplicateTimeException;
+import roomescape.reservation.application.exception.DeleteTimeException;
+import roomescape.reservation.application.exception.DuplicateTimeException;
 import roomescape.reservation.application.repository.ReservationRepository;
 import roomescape.reservation.application.repository.ReservationTimeRepository;
 import roomescape.reservation.presentation.dto.AvailableReservationTimeResponse;
-import roomescape.reservation.presentation.dto.ReservationResponse;
 import roomescape.reservation.presentation.dto.ReservationTimeRequest;
 import roomescape.reservation.presentation.dto.ReservationTimeResponse;
 
@@ -24,7 +23,7 @@ public class ReservationTimeService {
     }
 
     public ReservationTimeResponse createReservationTime(final ReservationTimeRequest reservationTimeRequest) {
-        if(reservationTimeRepository.isExists(reservationTimeRequest.getStartAt())){
+        if (reservationTimeRepository.isExists(reservationTimeRequest.getStartAt())) {
             throw new DuplicateTimeException("[ERROR] 중복된 시간은 추가할 수 없습니다.");
         }
 
@@ -38,7 +37,7 @@ public class ReservationTimeService {
     }
 
     public void deleteReservationTime(final Long id) {
-        if(reservationRepository.existsByTimeId(id)){
+        if (reservationRepository.existsByTimeId(id)) {
             throw new DeleteTimeException("[ERROR] 예약이 이미 존재하는 시간입니다.");
         }
         reservationTimeRepository.delete(id);
@@ -47,7 +46,8 @@ public class ReservationTimeService {
     public List<AvailableReservationTimeResponse> getReservationTimes(final LocalDate date, final Long themeId) {
         return reservationTimeRepository.findAllTimes().stream()
                 .map(reservationTime -> {
-                    boolean alreadyBooked = reservationRepository.existsByDateAndThemeIdAndTimeId(date, reservationTime.getId(), themeId);
+                    boolean alreadyBooked = reservationRepository.existsByDateAndThemeIdAndTimeId(date,
+                            reservationTime.getId(), themeId);
                     return new AvailableReservationTimeResponse(reservationTime, alreadyBooked);
                 })
                 .toList();
