+++ src/main/java/roomescape/reservation/infrastructure/dao/ReservationDao.java
@@ -11,10 +11,10 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
-import roomescape.global.exception.DeleteReservationException;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.Role;
 import roomescape.reservation.application.dto.CreateReservationRequest;
+import roomescape.reservation.application.exception.DeleteReservationException;
 import roomescape.reservation.application.repository.ReservationRepository;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationDate;
