+++ src/test/java/roomescape/auth/application/service/AuthServiceTest.java
@@ -6,10 +6,10 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
+import roomescape.auth.application.exception.InvalidMemberException;
 import roomescape.auth.presentation.JwtTokenProvider;
 import roomescape.auth.presentation.dto.LoginRequest;
 import roomescape.auth.presentation.dto.TokenResponse;
-import roomescape.global.exception.InvalidMemberException;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.Role;
 import roomescape.member.infrastructure.fake.FakeMemberDao;
