+++ src/test/java/roomescape/reservation/application/service/ReservationServiceTest.java
@@ -8,12 +8,12 @@ import java.time.LocalTime;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.global.exception.GetThemeException;
-import roomescape.global.exception.GetTimeException;
 import roomescape.member.application.repository.MemberRepository;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.Role;
 import roomescape.member.infrastructure.fake.FakeMemberDao;
+import roomescape.reservation.application.exception.GetThemeException;
+import roomescape.reservation.application.exception.GetTimeException;
 import roomescape.reservation.application.repository.ReservationRepository;
 import roomescape.reservation.application.repository.ReservationTimeRepository;
 import roomescape.reservation.application.repository.ThemeRepository;
