+++ src/test/java/roomescape/reservation/infrastructure/dao/ReservationDaoTest.java
@@ -12,12 +12,12 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.global.exception.DeleteReservationException;
 import roomescape.member.application.dto.CreateMemberRequest;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.Role;
 import roomescape.member.infrastructure.dao.MemberDao;
 import roomescape.reservation.application.dto.CreateReservationRequest;
+import roomescape.reservation.application.exception.DeleteReservationException;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationDate;
 import roomescape.reservation.domain.ReservationTime;
