+++ src/test/java/roomescape/reservation/infrastructure/dao/ReservationTimeDaoTest.java
@@ -10,8 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.global.exception.DeleteTimeException;
-import roomescape.global.exception.DuplicateTimeException;
+import roomescape.reservation.application.exception.DeleteTimeException;
 import roomescape.reservation.presentation.dto.ReservationTimeRequest;
 
 @JdbcTest
